Glossary: some edits to CDN entry
Previously, the entry said that the app developer uploaded assets to
the CDN. There are some CDNs that act that way, but more act as
caching proxies. In that model, the app developer sets up the CDN as
a proxy to their origin and uses the CDN URLs in their app.