Adding clarification on array.push()
I experienced much pain trying to find out why changing an attribute that was an array was not updating the #each helper result.

It is not clear from the documentation that such adding and removing needs to be done with pushObject. I hope this change will help save new Ember users some time.

If this is not the correct place to put such documentation, please let me know an alternative.