[BUGFIX] #399 Improve Naming Conventions table
The naming conventions table in older guides cuts off the text. And in the current guide making the text smaller is not very elegant.

This commit represents the convention as tree view. Which takes up more vertical space but is easier to visualize at a glance. And doesn't suffer the scroll overflow or font size issues.

I believe this should also be back ported to old versions of the guides. Because #399 was reported again v1.12.0 version of guides