a11y related tweaks to Templates -> Actions page
The first example on the Actions page contains an accessibility antipattern.
Essentially only users utilising a mouse can activate the 'click' event and subsequent action, whereas keyboard equivalents that normally also trigger a click will not work.
This breaks the default browser behaviour, which is why sticking to 'clickable' elements is best for a correctly working example.