Clarify wording on passing an action to a component
Currently the section headed "Passing the Action to the Component" begins with the sentence "Now we need to make it so that the `onConfirm()` event in the button-with-confirmation component triggers the `userDidDeleteAccount()` action..." However, since `onConfirm` hasn't yet been defined, the use of the definite article makes it challenging for a reader unfamiliar with the material to follow the thought process.

The proposed wording change avoids referring to `onConfirm` until after the example that introduces it. Instead we simply explain that we are going to pass the parent action to the component as a property. A note is also added to explicitly highlight the role that the `action` helper plays in the process.