update(testing): renames testing types
This renames testing types to avoid an arbitrary distinction between
application, integration and unit tests, but instead embrace a naming
which reflects which kind of test setup is needed to run them.
This introduces the testing type titles

- application test
- rendering test
- container test
- unit test

into the testing section as presented in @Turbo87 's talk "The Next
Generation of Testing", Dec 2017.