Update index.md
This cautionary note seems to be out of date. In a testing environment in a 3.0 app, after running ember s, typing `require('app-name/config/environment').default` in the console returns an object with `environment: "test"`. 

Also, in a 2.18 addon, `require('dummy/config/environment').default` returns an object with `environment: "test"`.